+++ src/api/controllers/user.controller.ts
@@ -1,8 +1,10 @@
 export {};
 import { NextFunction, Request, Response, Router } from 'express';
+const mongoose = require('mongoose');
+const ObjectId = mongoose.Types.ObjectId;
 const httpStatus = require('http-status');
 const { omit } = require('lodash');
-import { User } from 'api/models';
+import { User, UserNote } from 'api/models';
 import { startTimer, apiJson } from 'api/utils/Utils';
 const { handler: errorHandler } = require('../middlewares/error');
 
@@ -93,6 +95,24 @@ exports.list = async (req: Request, res: Response, next: NextFunction) => {
     startTimer(req);
     const data = (await User.list(req)).transform(req);
     apiJson({ req, res, data, model: User });
+  } catch (e) {
+    next(e);
+  }
+};
+
+/**
+ * Get user's notes.
+ * NOTE: Any logged in user can get a list of notes of any user.
+ * @public
+ * @example GET https://localhost:3009/v1/users/USERID/notes
+ */
+exports.listUserNotes = async (req: Request, res: Response, next: NextFunction) => {
+  try {
+    startTimer(req);
+    const userId = req.params.userId;
+    req.query = { ...req.query, user: new ObjectId(userId) }; // append to query (by userId) to final query
+    const data = (await UserNote.list({ query: req.query })).transform(req);
+    apiJson({ req, res, data, model: UserNote });
   } catch (e) {
     next(e);
   }