+++ src/api/models/index.ts
@@ -1,1 +1,2 @@
 export const User = require('./user.model');
+export const UserNote = require('./userNote.model');