+++ src/index.ts
@@ -34,6 +34,11 @@ server.listen(port, () => {
   console.log(`${mstime.end('app-start').last} ms`);
 });
 
+if (env === 'development') {
+  // initialize test data once (admin@example.com)
+  require('./api/utils/InitData');
+}
+
 /**
  * Exports express
  * @public